# gateway (copy)/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.GoalCreate:
    properties:
      CurrentAmount:
        type: number
      Deadline:
        type: string
      Name:
        type: string
      TargetAmount:
        type: number
      UserId:
        type: string
    type: object
  genproto.GoalGet:
    properties:
      CreatedAt:
        type: string
      CurrentAmount:
        type: number
      Deadline:
        type: string
      Id:
        type: string
      Name:
        type: string
      Status:
        type: string
      TargetAmount:
        type: number
      UpdatedAt:
        type: string
      UserId:
        type: string
    type: object
  genproto.GoalList:
    properties:
      Goals:
        items:
          $ref: '#/definitions/genproto.GoalGet'
        type: array
      Limit:
        type: integer
      Offset:
        type: integer
      TotalCount:
        type: integer
    type: object
...
  genproto.GoalUpt:
    properties:
      CurrentAmount:
        type: number
      Deadline:
        type: string
      Name:
        type: string
      Status:
        type: string
      TargetAmount:
        type: number
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
...
  /goals/{goal_id}:
    delete:
      consumes:
      - application/json
      description: Delete a goal by its ID
      parameters:
      - description: Goal ID
        in: path
        name: goal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goal deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Goal not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Goal
      tags:
      - Goal
    get:
      consumes:
      - application/json
      description: Retrieve a goal by its ID
      parameters:
      - description: Goal ID
        in: path
        name: goal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goal details
          schema:
            $ref: '#/definitions/genproto.GoalGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Goal not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Goal
      tags:
      - Goal
    put:
      consumes:
      - application/json
      description: Update an existing goal
      parameters:
      - description: Goal ID
        in: path
        name: goal_id
        required: true
        type: string
      - description: Goal data
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/genproto.GoalUpt'
      produces:
      - application/json
      responses:
        "200":
          description: Goal updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Goal
      tags:
      - Goal
  /goals/create:
    post:
      consumes:
      - application/json
      description: Create a new goal
      parameters:
      - description: Goal data
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/genproto.GoalCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Goal created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Goal
      tags:
      - Goal
  /goals/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of goals
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Goal Status
        in: query
        name: status
        type: string
      - description: Goal Name
        in: query
        name: name
        type: string
      - description: Target Amount From
        in: query
        name: target_from
        type: number
      - description: Target Amount To
        in: query
        name: target_to
        type: number
      - description: Deadline From
        in: query
        name: deadline_from
        type: string
      - description: Deadline To
        in: query
        name: deadline_to
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of goals
          schema:
            $ref: '#/definitions/genproto.GoalList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Goals
      tags:
      - Goal
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"